fibre visualization implemented. Not sure if it does this correctly.

+++ vtu_converter/generate_mesh.py
@@ -10,33 +10,52 @@ class mesh_generator:
 		self.refinement1 = refinement1
 		self.refinement2 = refinement2
 		self.refinement3 = refinement3	
+		self.refinement1_fibers = refinement1
+		self.refinement2_fibers = refinement2
+		self.refinement3_fibers = refinement3
 		self.out_file = out_file
-	def setRefinement(self, x):
-		self.refinement = x
-	def setExfile(self, exfile):
+	def set_refinement(self, x,y,z):
+		self.refinement1 = x
+		self.refinement2 = y
+		self.refinement3 = z
+	def set_refinement_fibers(self, x,y,z):
+		refinement1_fibers = x
+		refinement2_fibers = y
+		refinement3_fibers = z
+	def set_exfile(self, exfile):
 		self.exfile = exfile
-
+	def set_outfile(self, out_file):
+		self.out_file = out_file
 	def generate_mesh(self, delete_data = True):
 		ex_reader = exfile_reader.exfile_reader(self.exfile)
 		vtu_conv = vtu_converter.VTU_converter()
-		print(self.refinement1)
+		
 		if (self.refinement1 > 0 or self.refinement2 > 0 or self.refinement3 > 0):
 			ex_reader.kd_interpolation("interpolation.exfile", self.refinement1, self.refinement2, self.refinement3)
+			ex_reader.create_directional_field("fibres.exfile", self.refinement1, self.refinement2, self.refinement3)
+			ex_reader_fibres = exfile_reader.exfile_reader("fibres.exfile")
+			#ex_reader.set_file("interpolation.exfile")
 			ex_reader = exfile_reader.exfile_reader("interpolation.exfile")
+			ex_reader_fibres.create_node_coordinates("fibres_nodes.txt")
+		else:
+			ex_reader.create_directional_field("fibres.exfile", self.refinement1, self.refinement2, self.refinement3)
+			ex_reader_fibres = exfile_reader.exfile_reader("fibres.exfile")
+			ex_reader_fibres.create_node_coordinates("fibres_nodes.txt")
 		ex_reader.create_node_coordinates()
 		ex_reader.create_node_connectivity()
 		ex_reader.create_fibre_values()
 		prolate_spheroid_to_euclidian("coordinates_prolate_spheroidal.txt", "coordinates_euclidian.txt", 0.3525E+02)
+		prolate_spheroid_to_euclidian("fibres_nodes.txt", "fibres_nodes_euclidian.txt", 0.3525E+02)
 		coords = "coordinates_euclidian.txt"
 		cons = "connectivity.txt"
-		vtu_conv.directional_field( "coordinates_euclidian.txt", "fibre_values_spheroidal.txt","directional_field1.vtu")
+		vtu_conv.write_directional_field( "fibres_nodes_euclidian.txt","fibres.vtu", 3)
 
 		vtu_conv.heart_to_vtu(cons, coords, self.out_file)
-		"""if (delete_data == True):
+		if (delete_data == True):
 			directory = os.getcwd()
 			os.remove(os.path.join(directory, "connectivity.txt"))
 			os.remove(os.path.join(directory, "coordinates_euclidian.txt"))
-			os.remove(os.path.join(directory, "coordinates_prolate_spheroidal.txt"))"""
+			os.remove(os.path.join(directory, "coordinates_prolate_spheroidal.txt"))
 
-mesh_generator = mesh_generator("heart1.exfile", 1,1,1, "heart_123.vtu")#
+mesh_generator = mesh_generator("heart1.exfile", 2,2,2, "heart_111.vtu")#
 mesh_generator.generate_mesh()
